tweaking annotation

+++ Ramses_analysis/FU_ori_investigation/event_frames.py
@@ -157,7 +157,7 @@ for plot_time in plot_times:
         file = open(part_info_pickle, 'rb')
         part_info = pickle.load(file)
         file.close()
-    if len(part_info['particle_velocity']) != len(part_info['particle_tag']):
+    if len(part_info['particle_velocity']) != len(part_info['particle_tag']) and os.path.isfile(part_info_pickle):
         os.remove(part_info_pickle)
     #get relative velocity
     if os.path.isfile(part_info_pickle) == False:
